new cool patterns FactoryMethod 2nd example, Builder etc. - Styling improved

+++ index.php
@@ -1,11 +1,12 @@
 <?php
+
 /**
  * Created by PhpStorm.
  * User: Waldi
  * Date: 25.06.2017
  * Time: 22:05
  */
-declare(strict_types = 1);
+declare(strict_types=1);
 ?>
 <!doctype html>
 <html lang="en">
@@ -70,31 +71,35 @@ declare(strict_types = 1);
     $files = [];
     $dir = 'src/';
     $dh = opendir($dir);
-    echo '<ul class="list-group">';
-    while(false !== ($filename = readdir($dh))) {
-        if($filename !== '.' && $filename !== '..') {
+    $chosen = false;
+    echo '<ul>';
+    while (false !== ($filename = readdir($dh))) {
+        $desc = '';
+        if ($filename !== '.' && $filename !== '..') {
             $files[] = $filename;
-            echo '<a class="list-group-item list-group-item-action" href="?pattern='
-                .$filename.'"">'.
-                '<span class="list-pattern-entry btn btn-primary">'.$filename
-                .'</span>';
+            echo '<div class="row">                                       
+                    <div class="col-md-4">
+                     <a class="list-group-item list-group-item-action" href="?pattern=' . $filename . '"">
+                    <span class="list-pattern-entry btn btn-primary">' . $filename . '</span></a>
+                    </div>';
 
-            $desc = '';
-            if(file_exists($dir.$filename.'/readme.md')) {
-                $desc = file_get_contents($dir.$filename.'/readme.md');
-                echo ' &mdash; <span class="">'.$desc.'</span>';
+            echo '<div class="col-md-8 list-group-item">';
+            if (file_exists($dir . $filename . '/readme.md')) {
+                $desc = file_get_contents($dir . $filename . '/readme.md');
+                echo ' <span class="text-muted">' . $desc . '</span>';
             }
-            echo '</a>';
+            echo '</div>               
+               </div>';
         }
     }
     echo '</ul>';
 
-    echo '<span class="badge badge-pill badge-info">'.count($files)
-        .' patterns found </span>';
+    echo '<span class="badge badge-pill badge-info">' . count($files)
+        . ' patterns found </span>';
 
-    if(!empty($namespace = $_GET['pattern'])) {
+    if (array_key_exists('pattern', $_GET) && !empty($namespace = $_GET['pattern'])) {
         echo '<br><div class="card"><h4>Output</h4>';
-        $fullClass = 'Patterns\\'.$namespace.'\\Index';
+        $fullClass = 'Patterns\\' . $namespace . '\\Index';
         $pattern = new $fullClass;
         $pattern->start();
         echo '</div><br>';
@@ -106,33 +111,31 @@ declare(strict_types = 1);
         $content = highlight_file($filename, true);
         $file = explode('<br />', $content);
         $i = 1;
-        echo '<div class="card">'.$filename;
+        echo '<div class="card">' . $filename;
         echo '<div class="card-body">';
-        foreach($file as $line) {
-            if(empty($line)) {
+        foreach ($file as $line) {
+            if (empty($line)) {
                 $line = '&nbsp;';
             }
-            echo '<div class="syntax-highlight-line">'.str_pad(
-                    (string) $i,
+            echo '<div class="syntax-highlight-line">' . str_pad(
+                    (string)$i,
                     3,
                     '0',
                     STR_PAD_LEFT
                 )
-                .'</div> <div class="syntax-highlight-code">'.$line
-                .'</div>';
+                . '</div> <div class="syntax-highlight-code">' . $line
+                . '</div>';
             $i++;
         }
         echo '</div>';
         echo '</div>';
     }
 
-    //highlight_num('src/'.$namespace.'/Index.php');
-
     $listFiles = [];
-    if($chosen) {
-        foreach(glob('src/'.$namespace.'/*.php') as $filename) {
+    if ($chosen) {
+        foreach (glob('src/' . $namespace . '/*.php') as $filename) {
             //echo "$filename - Größe: ".filesize($filename)."\n";
-            if(stripos($filename, 'Index.php') !== false) {
+            if (stripos($filename, 'Index.php') !== false) {
                 array_unshift($listFiles, $filename);
                 continue;
             }
@@ -140,13 +143,10 @@ declare(strict_types = 1);
             $listFiles[] = $filename;
         }
 
-        foreach($listFiles as $filename) {
-            //highlight_num('src/'.$namespace.'/Index.php');
+        foreach ($listFiles as $filename) {
             highlight_num($filename);
         }
     }
-
-
     ?>
 </div>
